apps/api/requirements.txt
# FastAPI Backend Dependencies
# MedXrayChat API Server

# Core
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.18.3

# Redis (caching & sessions)
redis==7.1.0
aioredis==2.0.1

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0

# AI & Image Processing
ultralytics==8.4.9
torch==2.10.0
torchvision==0.25.0
transformers==5.0.0
accelerate==1.12.0
Pillow==12.1.0
opencv-python-headless==4.13.0.90
numpy==2.4.1
pydicom==3.0.1
grad-cam==1.5.4

# Qwen3-VL specific
qwen-vl-utils==0.0.14

# Utils
httpx==0.28.1
python-dotenv==1.2.1
loguru==0.7.3
aiofiles==25.1.0
slowapi==0.1.9

# Development
pytest==9.0.2
pytest-asyncio==1.3.0
black==26.1.0
ruff==0.14.14










